refactor header files
