+++ data/parse_data.h
@@ -54,128 +54,3 @@ private:
 
 };
 
-
-class ParseHealthData : public ParseData{
-  public:
-    class HealthData : public ParseData::Data {
-    public:
-          std::vector<int> id;
-          std::vector<float> lcavol;
-          std::vector<float>  lweight;
-          std::vector<int>  age;
-          std::vector<float>  lbph;
-          std::vector<float>  svi;
-          std::vector<float>  lcp;
-          std::vector<float>  gleason;
-          std::vector<float>  pgg45;
-          std::vector<float>  lpsa;
-          std::vector<std::string>  train;
-    };
-
-
-    Data read_txt(std::string input_file){
-      // Create a text string, which is used to output the text file
-      std::string myText;
-      HealthData output_data;
-      int line_num = 0;
-
-      // Read from the text file
-      std::ifstream MyReadFile(input_file);
-
-      // Use a while loop together with the getline() function to read the file line by line
-      while (getline (MyReadFile, myText)) {
-
-        std::stringstream line(myText);
-        int line_ind = 0;
-
-        if (line_num > 0){
-        while (line.good()) {
-          std::string substr;
-          std::getline(line, substr, ',');
-          switch(line_ind){
-          case 0:
-            output_data.id.push_back(std::stoi(substr));
-            break;
-          case 1:
-            output_data.lcavol.push_back(std::stof(substr));
-            break;
-          case 2:
-            output_data.lweight.push_back(std::stof(substr));
-            break;
-          case 3:
-            output_data.age.push_back(std::stof(substr));
-            break;
-          case 4:
-            output_data.lbph.push_back(std::stof(substr));
-            break;
-          case 5:
-            output_data.svi.push_back(std::stof(substr));
-            break;
-          case 6:
-            output_data.lcp.push_back(std::stof(substr));
-            break;
-          case 7:
-            output_data.gleason.push_back(std::stof(substr));
-            break;
-          case 8:
-            output_data.pgg45.push_back(std::stof(substr));
-            break;
-          case 9:
-            output_data.lpsa.push_back(std::stof(substr));
-            break;
-          case 10:
-            output_data.train.push_back(substr);
-            break;
-          }
-          //v.push_back(substr);
-          line_ind += 1;
-        }
-      }
-      line_num +=1;
-      }
-
-      for(std::string i : output_data.train)
-        std::cout << i << "\n";
-
-      // Close the file
-      MyReadFile.close();
-
-      return output_data;
-    }
-
-    Data read_csv(std::string input_file) {
-      std::cout << input_file;
-
-      // File pointer
-      std::fstream fin;
-
-//      std::cout << "Enter input file path "; // Type a number and press enter
-//      std::cin >> input_file;                // Get user input from the keyboard
-
-      // Open an existing file
-      fin.open(input_file, std::ios::in);
-
-      // Read the Data from the file
-      // as String Vector
-      std::string line, temp;
-
-      // Loop through each row
-      while (fin >> temp) {
-
-        // read an entire row and
-        // store it in a string variable 'line'
-        getline(fin, line);
-
-        // used for breaking words
-        std::stringstream s(line);
-
-        // Loop through each element of the line
-        while (s.good()) {
-          std::string element;
-          getline(s, element, ',');
-          std::cout << element;
-        }
-      }
-    }
-
-};
